+++ oxidauth-http/src/server/api/v1/auth/authenticate.rs
@@ -3,11 +3,12 @@ use serde::{Deserialize, Serialize};
 use tracing::info;
 use uuid::Uuid;
 
-use crate::{provider::Provider, response::Response};
 use oxidauth_kernel::auth::authenticate::{
     AuthenticateParams, AuthenticateService,
 };
 use oxidauth_kernel::error::IntoOxidAuthError;
+
+use crate::{provider::Provider, response::Response};
 
 pub type AuthenticateReq = AuthenticateParams;
