+++ oxidauth-http/src/server/api/v1/authorities/delete_authority.rs
@@ -2,10 +2,12 @@ use axum::{
     extract::{Path, State},
     response::IntoResponse,
 };
-use oxidauth_kernel::authorities::delete_authority::*;
-use oxidauth_kernel::error::IntoOxidAuthError;
-use oxidauth_permission::parse_and_validate;
 use serde::{Deserialize, Serialize};
+
+use oxidauth_kernel::{
+    authorities::delete_authority::*, error::IntoOxidAuthError,
+};
+use oxidauth_permission::parse_and_validate;
 use tracing::{info, warn};
 
 use crate::middleware::permission_extractor::{
