+++ oxidauth-rs/src/client/authorities/mod.rs
@@ -1,5 +1,6 @@
 pub mod create_authority;
 pub mod delete_authority;
+pub mod find_authority_by_client_key;
 pub mod find_authority_by_id;
 pub mod find_authority_by_strategy;
 pub mod list_all_authorities;
@@ -7,4 +8,4 @@ pub mod update_authority;
 
 pub use oxidauth_kernel::authorities::create_authority::*;
 
-use super::{fmt, handle_response, Client, Resource};
+use super::{Client, Resource, fmt, handle_response};
